% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_utils.R
\name{print.jobcode}
\alias{print.jobcode}
\title{Nice print .jobcode}
\usage{
\method{print}{jobcode}(x, ...)
}
\arguments{
\item{x}{Text to print}

\item{...}{Currently unused}
}
\description{
Nice print .jobcode
}
