% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job_utils.R
\name{export}
\alias{export}
\title{What to return from a job}
\usage{
export(value = "changed")
}
\arguments{
\item{value}{What to return. One of:
\itemize{
\item \code{"all"}: Return everything, including imports
\item \code{"changed"} (default): Return all variables that are not identical to import.
\item \code{"new"}: Return only new variable names.
\item \code{c(var1, var2, ...)}: Return these variable names.
\item \code{NULL} or \code{"none"}: Return nothing. This is particularly useful for unnamed code chunks.
}}
}
\value{
\code{NULL}
}
\description{
Call this function as the last line in \code{job::job()}. An error is raised if
it is called in another environment. Under the hood, this function merely
\code{rm()} variables that does not match \code{value}. Because \code{job::job()} returns
everything at the end of the script, this defines what is returned.
}
